requirements.txt
astropy
ipdb
numpy
pandas
python-dateutil
rasterio
scipy
scikit-learn
setuptools
tqdm
zipfile
git+https://github.com/DHI-GRAS/snap-graph
git+https://github.com/DHI-GRAS/creodias-finder
